Made home it's own screen
